Monster Nature - Interactive installation
We develop an Interactive installation where two users can interact. The
user #1 has a controller with some potentiometers and buttons, has the
role to create monster on the world to break the balance and make the
user#2 lose, each monster has its own weight and hability. The user #2
has a controller with an accelerometer and has the role to keep the
balances, depending on the weight distribution on the world.